ai-service/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# AI/ML Libraries
langchain==0.1.0
langchain-google-genai==0.0.6
google-generativeai==0.3.2

# Database dependencies
motor==3.3.2
pymongo==4.6.0
redis==5.0.1

# HTTP client
httpx==0.25.2
requests==2.31.0

# Authentication
PyJWT==2.8.0

# Utilities
python-dotenv==1.0.0
